+++ src/storage/utils/checksums/checksumWorker.js
