+++ src/storage/utils/fileinfo/v3.js
